requirements.txt
tensorflow
numpy
pillow
matplotlib
seaborn
tk
